+++ package.json
@@ -17,7 +17,7 @@
         "figlet": "^1.2.4",
         "handlebars": "^4.7.7",
         "inquirer": "^7.3.3",
-        "js-yaml": "^3.14.0",
+        "js-yaml": "^4.1.0",
         "lodash": "^4.17.21",
         "memorystream": "^0.3.1",
         "noble-ed25519": "^1.0.3",
@@ -39,7 +39,7 @@
         "@types/figlet": "^1.2.0",
         "@types/handlebars": "^4.1.0",
         "@types/inquirer": "^7.3.1",
-        "@types/js-yaml": "^3.12.5",
+        "@types/js-yaml": "^4.0.5",
         "@types/lodash": "^4.14.165",
         "@types/memorystream": "^0.3.0",
         "@types/mocha": "^8.2.2",
