+++ src/service/CertificateService.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { existsSync } from 'fs';
+import { existsSync, writeFileSync } from 'fs';
 import { join, resolve } from 'path';
 import { Convert, Crypto, NetworkType } from 'symbol-sdk';
 import { Logger } from '../logger';
@@ -26,6 +26,7 @@ import { Constants } from './Constants';
 import { FileSystemService } from './FileSystemService';
 import { RuntimeService } from './RuntimeService';
 import { Utils } from './Utils';
+import { YamlUtils } from './YamlUtils';
 
 export interface CertificateParams {
     readonly target: string;
@@ -145,7 +146,7 @@ export class CertificateService {
 
         BootstrapUtils.createDerFile(mainAccountPrivateKey, join(certFolder, 'ca.der'));
         BootstrapUtils.createDerFile(transportPrivateKey, join(certFolder, 'node.der'));
-        await BootstrapUtils.writeTextFile(
+        await YamlUtils.writeTextFile(
             join(certFolder, 'serial.dat'),
             (randomSerial?.trim() || Convert.uint8ToHex(Crypto.randomBytes(19))).toLowerCase() + '\n',
         );
@@ -155,7 +156,7 @@ export class CertificateService {
             presetData.caCertificateExpirationInDays,
             presetData.nodeCertificateExpirationInDays,
         );
-        await BootstrapUtils.writeTextFile(join(certFolder, 'createNodeCertificates.sh'), command);
+        await YamlUtils.writeTextFile(join(certFolder, 'createNodeCertificates.sh'), command);
 
         const { stdout, stderr } = await this.runOpenSslCommand(
             presetData.symbolServerImage,
@@ -187,7 +188,7 @@ export class CertificateService {
             transportPublicKey: providedCertificates.transport.publicKey,
             mainPublicKey: providedCertificates.main.publicKey,
         };
-        await BootstrapUtils.writeYaml(metadataFile, metadata, undefined);
+        await YamlUtils.writeYaml(metadataFile, metadata, undefined);
     }
 
     private async shouldGenerateCertificate(metadataFile: string, providedCertificates: NodeCertificates): Promise<boolean> {
@@ -195,7 +196,7 @@ export class CertificateService {
             return true;
         }
         try {
-            const metadata = BootstrapUtils.loadYaml(metadataFile, false) as CertificateMetadata;
+            const metadata = YamlUtils.loadYaml(metadataFile, false) as CertificateMetadata;
             return (
                 metadata.mainPublicKey !== providedCertificates.main.publicKey ||
                 metadata.transportPublicKey !== providedCertificates.transport.publicKey ||
@@ -346,5 +347,13 @@ echo "Certificate Created"
             const publicKey = extractKey(stdout.substring(stdout.indexOf(middle, index) + middle.length, stdout.indexOf(to, index)));
             return { privateKey: privateKey, publicKey: publicKey };
         });
+    }
+
+    public static createDerFile(privateKey: string, file: string): void {
+        writeFileSync(file, Convert.hexToUint8(this.toAns1(privateKey)));
+    }
+    public static toAns1(privateKey: string): string {
+        const prefix = '302e020100300506032b657004220420';
+        return `${prefix}${privateKey.toLowerCase()}`;
     }
 }
