+++ src/service/ConfigLoader.ts
@@ -145,7 +145,7 @@ export class ConfigLoader {
                     throw new Error(
                         `Nemesis distribution balance cannot be less than 0. Mosaic ${mosaicName}, distribution address: ${
                             d.address
-                        }, amount: ${d.amount}, index ${index}. \nDistributions are:\n${BootstrapUtils.toYaml(distribution)}`,
+                        }, amount: ${d.amount}, index ${index}. \nDistributions are:\n${YamlUtils.toYaml(distribution)}`,
                     );
                 }
                 return d.amount;
@@ -219,7 +219,7 @@ export class ConfigLoader {
                 throw new Error(
                     `Invalid nemgen total supplied value, expected ${
                         m.supply
-                    } but total is ${supplied}. \nDistributions are:\n${BootstrapUtils.toYaml(m.currencyDistributions)}`,
+                    } but total is ${supplied}. \nDistributions are:\n${YamlUtils.toYaml(m.currencyDistributions)}`,
                 );
             }
         });
